refactor: divide withContainer in several functions and extract container api class in its own file
Instead of setting up and tearing down a container for each test, the BeforeAll and AfterAll hooks can be used
to setup one container at the start, running all the tests against it and then tearing it down.
For this the setup phase of withContainer was extracted in its own function as well as the teardown phase.